reafacor: move out typeDefs, resolvers and datasources declarations from index.ts
